Extract SizeRatioLeafStrategy from SizeRatioCompactionStrategy
