+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/impl/SizeRatioCompactionStrategy.java
@@ -25,14 +25,12 @@ import sleeper.core.partition.Partition;
 import sleeper.statestore.FileInfo;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static sleeper.compaction.strategy.impl.CompactionUtils.getFilesInAscendingOrder;
 import static sleeper.configuration.properties.table.TableProperty.SIZE_RATIO_COMPACTION_STRATEGY_RATIO;
 
 /**
@@ -60,7 +58,7 @@ public class SizeRatioCompactionStrategy extends AbstractCompactionStrategy {
     private int maxConcurrentCompactionJobsPerPartition;
 
     public SizeRatioCompactionStrategy() {
-        super();
+        super(new SizeRatioLeafStrategy());
     }
 
     @Override
@@ -99,7 +97,7 @@ public class SizeRatioCompactionStrategy extends AbstractCompactionStrategy {
                 }
                 long maxNumberOfJobsToCreate = maxConcurrentCompactionJobsPerPartition - numConcurrentCompactionJobs;
                 LOGGER.info("Max jobs to create = {}", maxNumberOfJobsToCreate);
-                List<CompactionJob> jobs = createJobsForLeafPartition(partition, activeFilesWithNoJobId);
+                List<CompactionJob> jobs = leafStrategy.createJobsForLeafPartition(partition, activeFilesWithNoJobId);
                 LOGGER.info("Defined {} compaction job{} for partition {}", jobs.size(), 1 == jobs.size() ? "s" : "", partitionId);
                 while (jobs.size() > maxNumberOfJobsToCreate) {
                     jobs.remove(jobs.size() - 1);
@@ -122,74 +120,4 @@ public class SizeRatioCompactionStrategy extends AbstractCompactionStrategy {
                 .size();
     }
 
-    @Override
-    protected List<CompactionJob> createJobsForLeafPartition(Partition partition, List<FileInfo> fileInfos) {
-        // Find files that meet criteria, i.e. sum of file sizes excluding largest
-        // is >= ratio * largest file size.
-        List<FileInfo> filesThatMeetCriteria = getListOfFilesThatMeetsCriteria(partition, fileInfos);
-        if (null == filesThatMeetCriteria || filesThatMeetCriteria.isEmpty()) {
-            LOGGER.info("For partition {} there is no list of files that meet the criteria", partition.getId());
-            return Collections.EMPTY_LIST;
-        }
-        LOGGER.info("For partition {} there is a list of {} files that meet the criteria", partition.getId(), filesThatMeetCriteria.size());
-
-        // Iterate through these files, batching into groups of compactionFilesBatchSize
-        // and creating a job for each group as long as it meets the criteria.
-        List<CompactionJob> compactionJobs = new ArrayList<>();
-        if (filesThatMeetCriteria.size() <= compactionFilesBatchSize) {
-            compactionJobs.add(factory.createCompactionJob(filesThatMeetCriteria, partition.getId()));
-        } else {
-            int position = 0;
-            List<FileInfo> files = new ArrayList<>(filesThatMeetCriteria);
-            while (position < files.size()) {
-                List<FileInfo> filesForJob = new ArrayList<>();
-                int j;
-                for (j = 0; j < compactionFilesBatchSize && position + j < files.size(); j++) {
-                    filesForJob.add(files.get(position + j));
-                }
-                // Create job for these files if they meet criteria
-                List<Long> fileSizes = filesForJob.stream().map(FileInfo::getNumberOfRecords).collect(Collectors.toList());
-                if (meetsCriteria(fileSizes)) {
-                    LOGGER.info("Creating a job to compact {} files in partition {}",
-                            filesForJob.size(), partition.getId());
-                    compactionJobs.add(factory.createCompactionJob(filesForJob, partition.getId()));
-                    filesForJob.clear();
-                    position += j;
-                } else {
-                    position++;
-                }
-            }
-        }
-
-        return compactionJobs;
-    }
-
-    private List<FileInfo> getListOfFilesThatMeetsCriteria(Partition partition, List<FileInfo> fileInfos) {
-        List<FileInfo> filesInAscendingOrder = getFilesInAscendingOrder(partition, fileInfos);
-
-        while (filesInAscendingOrder.size() > 1) {
-            List<Long> fileSizes = filesInAscendingOrder.stream().map(FileInfo::getNumberOfRecords).collect(Collectors.toList());
-            if (meetsCriteria(fileSizes)) {
-                return filesInAscendingOrder;
-            } else {
-                filesInAscendingOrder.remove(filesInAscendingOrder.size() - 1);
-            }
-        }
-        return null;
-    }
-
-    private boolean meetsCriteria(List<Long> fileSizesInAscendingOrder) {
-        if (fileSizesInAscendingOrder.isEmpty() || 1 == fileSizesInAscendingOrder.size()) {
-            return false;
-        }
-        long largestFileSize = fileSizesInAscendingOrder.get(fileSizesInAscendingOrder.size() - 1);
-        LOGGER.info("Largest file size is {}", largestFileSize);
-        long sumOfOtherFileSizes = 0L;
-        for (int i = 0; i < fileSizesInAscendingOrder.size() - 1; i++) {
-            sumOfOtherFileSizes += fileSizesInAscendingOrder.get(i);
-        }
-        LOGGER.info("Sum of other file sizes is {}", sumOfOtherFileSizes);
-        LOGGER.info("Ratio * largestFileSize <= sumOfOtherFileSizes {}", (ratio * largestFileSize <= sumOfOtherFileSizes));
-        return ratio * largestFileSize <= sumOfOtherFileSizes;
-    }
 }
